Added ResponseEntity in responses of controllers, checking for the existence of records in the DB ('exists' methods in services)

+++ src/main/java/com/ouch/musicservice/web/AlbumController.java
@@ -3,7 +3,8 @@ package com.ouch.musicservice.web;
 import com.ouch.musicservice.model.Album;
 import com.ouch.musicservice.service.AlbumService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,28 +16,49 @@ public class AlbumController {
     private AlbumService albumService;
 
     @GetMapping
-    public List<Album> getAllAlbums() {
-        return albumService.getAllAlbums();
+    public ResponseEntity<List<Album>> getAllAlbums() {
+        List<Album> albums = albumService.getAllAlbums();
+        return new ResponseEntity<>(albums, HttpStatus.OK);
     }
 
     @GetMapping("/{albumId}")
-    public Album getAlbumById(@PathVariable Long albumId) {
-        return albumService.getAlbumById(albumId);
+    public ResponseEntity<Album> getAlbumById(@PathVariable Long albumId) {
+        Album album = albumService.getAlbumById(albumId);
+        if (album != null) {
+            return new ResponseEntity<Album>(album, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping
-    public void createNewAlbum(@RequestBody Album newAlbum) {
-        albumService.createNewAlbum(newAlbum);
+    public ResponseEntity<Void> createNewAlbum(@RequestBody Album newAlbum) {
+        if (!albumService.exists(newAlbum)) {
+            albumService.createNewAlbum(newAlbum);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
+        }
     }
 
     @PutMapping("/{albumId}")
-    public void updateAlbumById(@PathVariable Long albumId,
+    public ResponseEntity<Void> updateAlbumById(@PathVariable Long albumId,
                                 @RequestBody Album updatedAlbum) {
-        albumService.updateAlbumById(albumId, updatedAlbum);
+        if (albumService.existsById(albumId)) {
+            albumService.updateAlbumById(albumId, updatedAlbum);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{albumId}")
-    public void deleteAlbumById(@PathVariable Long albumId) {
-        albumService.deleteAlbumById(albumId);
+    public ResponseEntity<Void> deleteAlbumById(@PathVariable Long albumId) {
+        if (albumService.existsById(albumId)) {
+            albumService.deleteAlbumById(albumId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
